% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{md_body}
\alias{md_body}
\title{Create a dashboard body}
\usage{
md_body(...)
}
\arguments{
\item{...}{Items to put. Typically \code{\link{md_tab_items}}.}
}
\description{
Create a dashboard body
}
\examples{
if (interactive()) {
  library(shiny)
  library(shinymd)

  ui = md_page(
    header = md_header(),
    sidebar = md_sidebar(
      title = 'Dashboard Title',
      md_sidebar_menu(
        md_sidebar_menu_item(
          title = 'Tab 1',
          tab_name = 'tab_1',
          icon = md_icon('loyalty')
        ),
        md_sidebar_menu_item(
          title = 'Tab 2',
          tab_name = 'tab_2',
          icon = md_icon('face')
        )
      )
    ),
    body = md_body(
      md_tab_items(
        md_tab_item(
          tab_name = 'tab_1',
          tags$h2('Tab 1 content')
        ),
        md_tab_item(
          tab_name = 'tab_2',
          tags$h2('Tab 2 content')
        )
      )
    )
  )

  server = function(input, output, session) {}

  shinyApp(ui, server)
}
}
