% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{md_box}
\alias{md_box}
\title{Create a general purpose box}
\usage{
md_box(..., width = 6, status = "success", title = NULL,
  subtitle = NULL)
}
\arguments{
\item{...}{Items to put.}

\item{width}{Width of the box.}

\item{status}{Color of the box header.}

\item{title}{Title to display.}

\item{subtitle}{Subtitle to display below title.}
}
\description{
Create a general purpose box
}
