% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{md_chart_box}
\alias{md_chart_box}
\title{Create a chart box}
\usage{
md_chart_box(..., id = NULL, width = 6, status = "success",
  title = "", subtitle = "", footer = NULL)
}
\arguments{
\item{...}{Items to put.}

\item{id}{Output variable name.}

\item{width}{Width of the box.}

\item{status}{Color of the chart area.}

\item{title}{Title to display.}

\item{subtitle}{Subtitle to display below title.}

\item{footer}{A footer to include in the box. Should be produced by
\code{\link{md_footer}}.}
}
\description{
Create a chart box
}
