% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\name{md_dropdown_menu}
\alias{md_dropdown_menu}
\title{Create a dropdown menu}
\usage{
md_dropdown_menu(..., icon = md_icon("notifications"))
}
\arguments{
\item{...}{Items to put. Typically, \code{\link{md_dropdown_menu_item}}.}

\item{icon}{Icon to identify the dropdown menu. Should be produced by
\code{\link{md_icon}}.}
}
\description{
Create a dropdown menu
}
\seealso{
\code{\link{md_header}} for example usage.
}
