% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\name{md_dropdown_menu_item}
\alias{md_dropdown_menu_item}
\title{Create a dropdown menu item}
\usage{
md_dropdown_menu_item(message)
}
\arguments{
\item{message}{Message to display.}
}
\description{
Create a dropdown menu item
}
\seealso{
\code{\link{md_header}} for example usage.
}
