% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\name{md_dropdown_menu_output}
\alias{md_dropdown_menu_output}
\title{Create a dropdown menu programmatically (UI-side)}
\usage{
md_dropdown_menu_output(id)
}
\arguments{
\item{id}{Output variable name.}
}
\description{
Create a dropdown menu programmatically (UI-side)
}
\seealso{
\code{\link{md_header}} for example usage and \code{\link{md_render_dropdown_menu}}
    for corresponding server-side function.
}
