% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{md_footer}
\alias{md_footer}
\title{Create a footer}
\usage{
md_footer(text, icon)
}
\arguments{
\item{text}{Text to display.}

\item{icon}{Icon to display alongside text. Should be produced by
\code{\link{md_icon}}.}
}
\description{
Create a footer
}
