% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\name{md_header}
\alias{md_header}
\title{Create a dashboard header}
\usage{
md_header(...)
}
\arguments{
\item{...}{Items to put in the dashboard header.}
}
\description{
Create a dashboard header
}
\examples{
if (interactive()) {
  library(shiny)
  library(shinymd)

  ui = md_page(
    header = md_header(
      md_dropdown_menu(
        icon = md_icon('shopping_cart'),
        md_dropdown_menu_item('Message 1'),
        md_dropdown_menu_item('Message 2')
      ),

      # Create dropdown menu programmatically (UI-side)
      md_dropdown_menu_output('dropdown_menu')
    ),
    sidebar = md_sidebar(),
    body = md_body()
  )

  server = function(input, output) {
    # Create dropdown menu programmatically (server-side)
    output$dropdown_menu = md_render_dropdown_menu({

      # Assuming the dropdown menu's message stored in a list
      notif = list(
        'Message 3',
        'Message 4',
        'Message 5'
      )

      items = lapply(notif, function(x) {
        md_dropdown_menu_item(x)
      })

      md_dropdown_menu(
        icon = md_icon('notifications'),
        items
      )
    })
  }

  shinyApp(ui, server)
}
}
