% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{md_icon}
\alias{md_icon}
\title{Create a Material icon}
\usage{
md_icon(name)
}
\arguments{
\item{name}{Name of the icon. See \url{https://design.google.com/icons/}
for available icons.}
}
\description{
Create a Material icon
}
