% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{md_info_box}
\alias{md_info_box}
\title{Create an info box}
\usage{
md_info_box(value, width = 3, status = "info", title,
  title_icon = NULL, footer = NULL)
}
\arguments{
\item{value}{Value to display. Usually a number or short text.}

\item{width}{Width of the box.}

\item{status}{Main color of the box.}

\item{title}{Title to display.}

\item{title_icon}{Icon to display alongside title. Should be produced by
\code{\link{md_icon}}.}

\item{footer}{A footer to include in the box. Should be produced by
\code{\link{md_footer}}.}
}
\description{
Create an info box
}
