% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{md_info_box_output}
\alias{md_info_box_output}
\title{Create an info box programmatically (UI-side)}
\usage{
md_info_box_output(id, width = 3)
}
\arguments{
\item{id}{Output variable name.}

\item{width}{Width of the box.}
}
\description{
Create an info box programmatically (UI-side)
}
