% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{md_page}
\alias{md_page}
\title{Create dashboard page}
\usage{
md_page(title = "My Dashboard", header = NULL, sidebar = NULL,
  body = NULL)
}
\arguments{
\item{title}{Title to display in browser's title bar.}

\item{header}{Dashboard header produced by \code{\link{md_header}}.}

\item{sidebar}{Dashboard sidebar produced by \code{\link{md_sidebar}}.}

\item{body}{Dashboard body produced by \code{\link{md_body}}.}
}
\description{
Create dashboard page
}
\examples{
if (interactive()) {
  library(shiny)
  library(shinymd)

  ui = md_page(
    header = md_header(),
    sidebar = md_sidebar(),
    body = md_body()
  )

  server = function(input, output) {}

  shinyApp(ui, server)
}
}
