% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\name{md_render_dropdown_menu}
\alias{md_render_dropdown_menu}
\title{Create a dropdown menu programmatically (server-side)}
\usage{
md_render_dropdown_menu(expr)
}
\arguments{
\item{expr}{Expression that returns dropdown menu with its items.}
}
\description{
Create a dropdown menu programmatically (server-side)
}
\seealso{
\code{\link{md_header}} for example usage and \code{\link{md_dropdown_menu_output}}
    for corresponding UI-side function.
}
