% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{md_render_info_box}
\alias{md_render_info_box}
\title{Create an info box programmatically (server-side)}
\usage{
md_render_info_box(expr)
}
\arguments{
\item{expr}{Expression that returns an info box.}
}
\description{
Create an info box programmatically (server-side)
}
