% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{md_sidebar}
\alias{md_sidebar}
\title{Create dashboard sidebar}
\usage{
md_sidebar(..., title = "My Dashboard", color = "green",
  background_color = "white", background_image = TRUE,
  title_url = NULL)
}
\arguments{
\item{...}{Items to put in the dashboard sidebar.}

\item{title}{Title to display in the top side of the sidebar.}

\item{color}{Sidebar color theme.}

\item{background_color}{Sidebar's background color.}

\item{background_image}{Sidebar's background image.}

\item{title_url}{Title's redirect URL.}
}
\description{
Create dashboard sidebar
}
\examples{
if (interactive()) {
  library(shiny)
  library(shinymd)

  ui = md_page(
    header = md_header(),
    sidebar = md_sidebar(
      title = 'Dashboard Title',
      md_sidebar_menu(
        md_sidebar_menu_item(
          title = 'Tab 1',
          tab_name = 'tab_1',
          icon = md_icon('loyalty')
        ),
        md_sidebar_menu_item(
          title = 'Tab 2',
          tab_name = 'tab_2',
          icon = md_icon('face')
        )
      )
    ),
    body = md_body()
  )

  server = function(input, output, session) {}

  shinyApp(ui, server)
}
}
