% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{md_sidebar_menu}
\alias{md_sidebar_menu}
\title{Create a sidebar menu}
\usage{
md_sidebar_menu(...)
}
\arguments{
\item{...}{Items to put. Typically \code{\link{md_sidebar_menu_item}}s.}
}
\description{
Create a sidebar menu
}
\seealso{
\code{\link{md_sidebar}} for example usage.
}
