% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{md_sidebar_menu_item}
\alias{md_sidebar_menu_item}
\title{Create a sidebar menu item.}
\usage{
md_sidebar_menu_item(tab_name, title, icon = md_icon("dashboard"))
}
\arguments{
\item{tab_name}{Identifier that links menu item with \code{\link{md_tab_item}}.
Should correspond exactly to the \code{tab_name} argument given in
\code{\link{md_tab_item}}.}

\item{title}{Title to identify the menu item.}

\item{icon}{Icon to identify the menu item.}
}
\description{
Create a sidebar menu item.
}
\seealso{
\code{\link{md_sidebar}} for example usage.
}
