% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{md_tab_box}
\alias{md_tab_box}
\title{Create a tabbed box}
\usage{
md_tab_box(..., id = NULL, title = NULL, width = 6,
  status = "success")
}
\arguments{
\item{...}{Items to put. Should be \code{\link{md_tab_panel}}s.}

\item{id}{Variable output name.}

\item{title}{Title to display.}

\item{width}{Width of the box.}

\item{status}{Color of the box header.}
}
\description{
Create a tabbed box
}
