% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{md_tab_item}
\alias{md_tab_item}
\title{Create a tab}
\usage{
md_tab_item(tab_name, ...)
}
\arguments{
\item{tab_name}{Identifier that links tab with \code{\link{md_sidebar_menu_item}}.
Should correspond exactly to the \code{tab_name} argument given in
\code{\link{md_sidebar_menu_item}}}

\item{...}{Items to put.}
}
\description{
Create a tab
}
\seealso{
\code{\link{md_body}} for example usage.
}
