% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{md_tab_items}
\alias{md_tab_items}
\title{Create a tabs container}
\usage{
md_tab_items(...)
}
\arguments{
\item{...}{Items to put. Typically \code{\link{md_tab_item}}.}
}
\description{
Create a tabs container
}
\seealso{
\code{\link{md_body}} for example usage.
}
