% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{md_tab_panel}
\alias{md_tab_panel}
\title{Create a tab panel}
\usage{
md_tab_panel(..., title, icon = NULL)
}
\arguments{
\item{...}{Items to put.}

\item{title}{Title to display.}

\item{icon}{Icon to display alongside title. Should be produced by
\code{\link{md_icon}}}
}
\description{
Create a tab panel that can be included within a \code{\link{md_tab_box}}
}
